add account page, and cart page

+++ app/src/main/java/com/amijul/mystore/data/remote/ProductFirestoreDataSource.kt
@@ -29,7 +29,7 @@ class ProductFirestoreDataSource(
             ProductUiModel(
                 id = id,
                 name = name,
-                price = priceNumber,
+                price = priceNumber.toFloat(),
                 unit = unit,
                 imageUrl = imageUrl,
                 inStock = inStock
